requirements.txt
requests
beautifulsoup4
schedule
python-dotenv
pytz
urllib3